Split library into packages
It's becoming necessary before adding more stuff and changing everything all over.
First level, i.e. neurom/__init__.py does not expose anything directly to force user to use the paths.

Ref #4.
