+++ src/neurom/__init__.py
@@ -1,8 +1,1 @@
-from .elements import *
-from .shape_functions import *
-from .quadratures import *
-from .mesh import *
-from .field import *
-from .interpolator import *
-from .integrator import *
-from .fem_model import *
+__version__ = "0.1.0"
