+++ src/neurom/geometry/__init__.py
@@ -1,0 +1,2 @@
+from neurom.geometry.iso_parametric_mapping_1d import IsoparametricMapping1D
+from neurom.geometry.barycentric_to_reference import barycentric_to_reference
