+++ src/neurom/meshes/__init__.py
@@ -1,0 +1,2 @@
+from neurom.meshes.topology import Topology
+from neurom.meshes.mesh import Mesh
