+++ src/neurom/reference_elements/__init__.py
@@ -1,0 +1,2 @@
+from neurom.reference_elements.reference_element import ReferenceElement
+from neurom.reference_elements.segment import Segment
