+++ tests/unit/test_iso_parametric_1d.py
@@ -1,4 +1,3 @@
-from abc import ABC, abstractmethod
 import pytest
 import torch
 import torch.nn as nn
